requirements.txt
numpy
opencv-python
pygame
Pillow
PyAutoGUI
future
pynput
autoit
